Create number-guessing-game.py
first attempt at a simple number guessing game in Python. This program has the user attempt to guess a number between 1-100. Each time the user guesses, the program will tell the user if they were too high or too low. If they guess correctly, they are told so and the program ends.