v1.2.7 | Perfect Model v1 added | (Update) number-guessing-game.py
Added 'Perfect' AI model. Each time the AI guesses, both the random number and guess will be random. This means that the AI must guess correctly on it's first and only attempt. The AI will repeat this 1000 times. Then, it will list the number of times it was able to guess the number first try from those attempts. It will then update a high score counter for the most correct guesses in 1000 attempts. There is also stats for number of guesses total, and wins (which is the same as perfect guesses because the perfect model can only win a perfect game). All of this is written in the perfect AI model function. Currently, there seems to be some issues with the stats updating and displaying on the AI stats screen. It's also worth noting that the variables for some of the perfect models stats have been changed, meaning they are out of order for other AI model stats. Also remembered to update versions variable.